refactor: split control plane task approval mapping
